requirements.txt
pandas 
psycopg2
openai
python-dotenv